cartChanges

+++ cart.php
@@ -4,33 +4,92 @@ class Cart{
     private $cartID;
     private $cartTotal;
     
-    public function __construct($userID){
+    public function __construct($userID)
+	{
         $this->itemsInCart = [];
-        $this->cartID = 0;
         $this->cartTotal = 0;
+		require_once 'login.php';
+        $conn = new mysqli($hn, $un, $pw, $db);
+        if($conn->connect_error)
+            die($conn->connect_error);
+        $query = "INSERT INTO cart VALUES (".$userID.")";
+		$result = $conn ->query($query);
+		if(!result) die(conn->connection_error);
+		$query1 = "SELECT cartID FROM cart WHERE userID = ".$userID;
+		$cartID = $conn ->query($query1);
+		if(!cartID) die(conn->connection_error);
     }
     
-    public function deleteCart(userID){
+    public function deleteCart(userID)
+	{
         require_once 'login.php';
         $conn = new mysqli($hn, $un, $pw, $db);
         if($conn->connect_error)
             die($conn->connect_error);
-        $query = "DELETE FROM cart WHERE userID =". 
-            $userID;
-        
+		
+		$query0 = "SELECT isAdmin FROM user WHERE userID = ".$userID;
+		$isAdmin = $conn ->query($query0);
+		if(!isAdmin) die(conn->connection_error);
+		
+		$query1 = "SELECT userID FROM cart WHERE cartID = ".$cartID;
+		$isCorrectUser = $conn ->query($query1);
+		if(!isCorrectUser) die(conn->connection_error);
+		
+		if($isAdmin)
+		{
+			$query = "DELETE FROM cart WHERE userID =". 
+				$userID;
+		}
+		else
+		{
+			if($isCorrectUser == $userID)
+			{
+				$query = "DELETE FROM cart WHERE userID =". 
+				$userID;
+			}
+		}
+		$result = $conn ->query($query);
+		if(!result) die(conn->connection_error);
     }
     
-    public function getItemsInCart(){
+    public function getItemsInCart()
+	{
         return($this->itemsInCart);
     }
     
-    public function getCartID(){
+    public function getCartID()
+	{
         return($this->itemsInCart);
     }
     
-    public function getCartTotal(){
+    public function getCartTotal()
+	{
         return($this->cartTotal);
     }
+	
+	public function deleteItemsFromCart ()
+	{
+		require_once 'login.php';
+        $conn = new mysqli($hn, $un, $pw, $db);
+        if($conn->connect_error)
+            die($conn->connect_error);
+        $query = "DELETE FROM cart_items WHERE itemID =". 
+            $itemID;
+		$result = $conn ->query($query);
+		if(!result) die(conn->connection_error);
+	}
+	
+	public function addItemsToCart ()
+	{
+		require_once 'login.php';
+        $conn = new mysqli($hn, $un, $pw, $db);
+        if($conn->connect_error)
+            die($conn->connect_error);
+        $query = "INSERT INTO cart_items VALUES (".$cartID.", ".$userID.", 
+		         ".$itemID.", 1".$priceTotal.")";
+		$result = $conn ->query($query);
+		if(!result) die(conn->connection_error);
+	}
 }
